% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_exclusion.R
\name{add_exclusion}
\alias{add_exclusion}
\title{Add the exclustion criteria to Phecode Table
Add NA to exclude those who cannot be regarded as controls for each phecode}
\usage{
add_exclusion(pheno_dat)
}
\arguments{
\item{pheno_data}{Phenotype data(sparse matrix)}
}
\description{
Add the exclustion criteria to Phecode Table
Add NA to exclude those who cannot be regarded as controls for each phecode
}
\examples{
add_exclusion(pheno_dat)
}
