% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas.R
\name{build_formula}
\alias{build_formula}
\title{Build formula for EnWAS}
\usage{
build_formula(base_model, exposure_var, inv = FALSE)
}
\arguments{
\item{base_model}{the formula of the model with the covariates}

\item{exposure_var}{exposure variables (aka. phenotype lists or predictors)}

\item{inv}{Boolean flag for use inverse normal transformation}
}
\value{
formula for EnWAS
}
\description{
Build formula for EnWAS
}
\examples{
build_formula("BMXWAIST ~ RIDAGEYR*RIAGENDR + BMXBMI", phenotype_lists)
}
