% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_restrictPhecodesBySex.R
\name{db_restrictPhecodesBySex}
\alias{db_restrictPhecodesBySex}
\title{db_restrictPhecodesBySex}
\usage{
db_restrictPhecodesBySex(phenotypes, id.sex)
}
\arguments{
\item{phenotypes}{The PheWAS table to have restrictions applied. The first column should be the id.}

\item{id.sex}{A \code{tbl()} source connection with the first column being the id and the second the gender, "M" or "F", of the
individual. Individuals with any other specification will have all gender specific phenotypes set to NA.}
}
\value{
The \code{phenotypes} \code{tbl()} source connection with NA values for individuals that do not match the gender for
gender-specific codes.
}
\description{
\code{restrictPhecodesBySex} alters a table for PheWAS with phecodes, as from \code{\link[PheWAS:createPhewasTable]{createPhewasTable}},
to exclude individuals with non-applicable genders from certain phenotypes.
}
\details{
Add PheWAS code descriptions to existing data.
}
\keyword{utilities}
