% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_code_agg.R
\name{default_code_agg}
\alias{default_code_agg}
\title{Sum or count distinct}
\usage{
default_code_agg(index)
}
\arguments{
\item{index}{an index}
}
\value{
the sum of an index or a length of an index
}
\description{
a function that either returns the sum of an input, if numeric, or the length
of an input if not numeric
}
