% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ.R
\name{.onAttach}
\alias{.onAttach}
\title{Startup Message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{N/A}

\item{pkgname}{N/A}
}
\value{
A start up message
}
\description{
Startup Message
}
\examples{
n/a
}
