% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas.R
\name{enwas_cv}
\alias{enwas_cv}
\title{Environment‐Wide Association Study (EnWAS) with testing}
\usage{
enwas_cv(
  base_model,
  exposure_vars,
  train_set,
  test_set,
  lab_col = "DIASTOLIC",
  inv_norm = FALSE
)
}
\arguments{
\item{base_model}{a string of base model}

\item{exposure_vars}{the phenotype list}

\item{train_set}{taring set}

\item{test_set}{test set}

\item{lab_col}{label column}

\item{inv_norm}{}
}
\value{
the model list and EnWAS result
}
\description{
Environment‐Wide Association Study (EnWAS) with testing
}
\examples{
linear_model <- 'BMXWAIST ~ RIDAGEYR*RIAGENDR + BMXBMI'
linear_res <- enwas_cv(linear_model, exposure_vars, train_set,test_set)
}
