% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_phewas}
\alias{example_phewas}
\title{example_phewas}
\format{
\subsection{\code{example_phewas}}{

data frame with 1825 rows and 17 columns
\describe{
\item{phenotype}{Sample Phenotypes}
\item{snp}{sample SNP}
\item{covariates}{Sample Covariates}
\item{beta}{the beta of the sample analysis}
\item{SE}{Standard Error of the sample analysis}
\item{OR}{Odds Ratio of the sample analysis}
\item{p}{P value of the sample analysis}
\item{type}{Logistic Regression}
\item{n_total}{number of samples in sample analysis}
\item{n_cases}{number of cases in sample analysis}
\item{n_controls}{number of controls in sample analysis}
\item{HWE_p}{Hardy-Weinberg}
\item{allele_freq}{Allele Frequency}
\item{n_no_snp}{Fill In Later}
\item{formula}{Fill In Later}
\item{expanded_formula_note}{Fill in Later}
...
}
}
}
\usage{
example_phewas
}
\description{
Example Data used in documentation
}
\keyword{datasets}
