% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest Plot for single model}
\usage{
forest_plot(xwas_result, top_n = 20)
}
\arguments{
\item{xwas_result}{the EnWAS result data frame}

\item{top_n}{the top to n phenotype}
}
\value{
plot result
}
\description{
The top n records are selected by the absolute values of the estimators.
}
\examples{
forest_plot(linear_res$enwas_res)
}
