% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{forest_plot_mult}
\alias{forest_plot_mult}
\title{Plot Forest for Multiple Models}
\usage{
forest_plot_mult(xwas_result_list, top_n = 20)
}
\arguments{
\item{xwas_result_list}{list of the EnWAS result data frames}

\item{top_n}{select top n record by the differences of the estimates}
}
\value{
plot multiple model results in a single image
}
\description{
Plot Forest for Multiple Models
}
\examples{
forest_plot_mult(list(linear = lm_enwas$enwas_res,
                               ns = en_enwas$enwas_res))
}
