% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{g_raw}
\alias{g_raw}
\title{Plot the outcomes against terms}
\usage{
g_raw(
  data,
  x = "BMXBMI",
  y = "DIASTOLIC",
  gender = "RIAGENDR",
  xlab = "BMI (kg/m²)",
  ylab = "Diastolic (mm Hg)",
  sample_ratio = 0.3
)
}
\arguments{
\item{data}{data}

\item{x}{terms or phenotype}

\item{y}{outcomes}

\item{gender}{gender of the cohort for plot point color}

\item{xlab}{x label text}

\item{ylab}{y label text}

\item{sample_ratio}{sample points to plot points, plot all the points is not helpful and time consuming when data set is large}
}
\value{
plot results
}
\description{
Plot the outcomes against terms
}
\examples{
g_raw(nhanes)
g_raw(nhanes,x="RIDAGEYR",xlab="Age (Years)")
}
