% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_restriction}
\alias{gender_restriction}
\title{Gender Restriction Table}
\format{
\subsection{\code{gender_restriction}}{

data frame with 1817 rows and 3 columns
\describe{
\item{phecode}{The phecodes}
\item{male_only}{a boolean indicating whether the phecode is male only}
\item{female_only}{a boolean indicating whether the phecode is female only}
...
}
}
}
\usage{
gender_restriction
}
\description{
A data map that connects phecodes to their gender tendencies.
}
\keyword{datasets}
