% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-enwas.R
\name{generic_enwas}
\alias{generic_enwas}
\title{Generic EnWAS}
\usage{
generic_enwas(data, base_model, expvars, ..., useResiduals = FALSE)
}
\arguments{
\item{data}{data.frame that the base_model was fit to and all exposures}

\item{base_model}{character representation of the base_model}

\item{expvars}{vector of the names of the exposures to be analyzed}

\item{useResiduals}{logical vector indicating whether or not to fit the model to the residuals from the base_model fit}
}
\value{
A data.frame where each row consists of a set of summary statistics for each exposure variable.
 The estimates from the model, upper and lower confidence bounds (UCL and LCL), BH adjusted p-values and
the multiple Rsquared value from the regression that was performed for that variable (either whole model or
against the residuals).
}
\description{
Generic EnWAS
}
\details{
This function examines the effects of the exposures listed in `expvars` in a model that is described by the `base_model`.
It uses an internal function to perform per variable regressions and returns a data.frame with summary statistics for 
each of those fits. If `useResiduals=TRUE` then the base model is fit, and the residuals obtained, each regression is
then against those residuals. If `useResiduals=FALSE` then for each exposure the model is augmented by that exposure and
fit.
}
