% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pheno.R
\name{get_pheno}
\alias{get_pheno}
\title{Get Pheno is a helper function for the long table PheWAS format}
\usage{
get_pheno(pheno, demos, phecode, MCC = 2)
}
\arguments{
\item{pheno}{Phecode Counts table}

\item{demos}{Demos table - mainly requires sex}

\item{phecode}{Phecode info table}

\item{MCC}{Minimum code count, default is 2}
}
\value{
a data table containing the demographic data with an additional column
named 'pheno' where 1 is case, 0 is control, -9 is excluded
}
\description{
Get Pheno is a helper function for the long table PheWAS format
}
