% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{lollipop}
\alias{lollipop}
\title{Plot for EnWAS AC/AC}
\usage{
lollipop(data, y = "AIC", x = "terms", top_n = 20, is_desc = FALSE)
}
\arguments{
\item{data}{AC/AC matrix}

\item{y}{metrics}

\item{x}{term}

\item{top_n}{top n terms}

\item{is_desc}{flag to decide desc}
}
\value{
plot results
}
\description{
Plot for EnWAS AC/AC
}
\examples{
lollipop(ns_enwas$qc_mtx,y="p_LRT")
}
