% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching}
\alias{matching}
\title{Propensity score matching to get matched controls}
\usage{
matching(pheno_dat, demo_dat, status_dat, ratio)
}
\arguments{
\item{pheno_dat}{Phenotype data(sparse matrix)}

\item{demo_dat}{Demographics dataframe with id and other covariates}

\item{status_dat}{A dataframe with id and case/control status}

\item{ratio}{1:ratio matched controls}
}
\description{
Propensity score matching to get matched controls
}
