% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe_as_ext.R
\name{phe_as_ext}
\alias{phe_as_ext}
\title{Function to perform and return GLM results on a single genotype-phenotype
association.}
\usage{
phe_as_ext(
  phe.gen,
  additive.genotypes = T,
  min.records = 20,
  return.models = F,
  confint.level = NA,
  factor.contrasts = NA,
  my.data,
  ...
)
}
\arguments{
\item{phe.gen}{A list with the name of the phecode column being used in the
regression, the column name of independent variable, and covariates}

\item{additive.genotypes}{Are additive genotypes being supplied? If so,
it will attempt to calculate allele frequencies and HWE values. Default
is TRUE.}

\item{min.records}{The minimum number of records to perform a test. For
logistic regression, there must be at least this number of each cases and
controls, for linear regression this total number of records. Default is 20}

\item{return.models}{Return a list the complete models, with the names equal
to the string formula used to create them, as well as the results. Default is
FALSE.}

\item{confint.level}{Uses the \code{MASS} package and the \code{confint}
function to calculate a confidence interval at the specified level.
\code{confint} uses a profile likelihood method, which takes some time to
compute. Output is stored in the \code{lower} and \code{upper} columns.
Logistic models will report OR CIs and linear models will report beta CIs.
Default is NA, which does not calculate confidence intervals.}

\item{factor.contrasts}{Contrasts used for factors to generate names used
in clogit.}

\item{my.data}{The data table containing the data for the regression.}

\item{...}{N/A}
}
\value{
The following are the default columns included in the returned data
frame. The attributes of the returned data frame contain additional
information about the anaylsis. This returns an output withh a single row for
a single genotype-phenotyping analysis
\item{phenotype}{The outcome under study}
\item{snp}{The predictor under study}
\item{adjustment}{The one off adjustment used}
\item{beta}{The beta coefficient for the predictor}
\item{SE}{The standard error for the beta coefficient}
\item{lower.p}{The lower bound of the quick confidence interval, if requested}
\item{upper.p}{The upper bound of the quick confidence interval, if requested}
\item{lower}{The lower bound of the \code{confint} confidence interval, if
requested}
\item{upper}{The upper bound of the \code{confint} confidence interval, if
requested}
\item{OR}{For logistic regression, the odds ratio for the predictor}
\item{p}{The p-value for the predictor}
\item{type}{The type of regression model used}
\item{n_total}{The total number of records in the analysis}
\item{n_cases}{The number of cases in the analysis (logical outcome only)}
\item{n_controls}{The number of controls in the analysis (logical outcome only)}
\item{HWE_p}{The Hardy-Weinberg equilibrium p-value for the predictor,
assuming 0,1,2 allele coding}
\item{allele_freq}{The allele frequency in the predictor for the coded allele}
\item{n_no_snp}{The number of records with a missing predictor}
}
\description{
Function to perform and return GLM results on a single genotype-phenotype
association.
}
