% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe_as_logistf.R
\name{phe_as_logistf}
\alias{phe_as_logistf}
\title{Function to perform and return logistf results on a single genotype-phenotype
association.}
\usage{
phe_as_logistf(
  phe.gen,
  additive.genotypes = T,
  min.records = 20,
  return.models = F,
  confint.level = NA,
  factor.contrasts = NA,
  my.data,
  ...
)
}
\arguments{
\item{phe.gen}{A list with the name of the phecode column being used in the
regression, the column name of independent variable, and covariates}

\item{additive.genotypes}{Are additive genotypes being supplied? If so,
it will attempt to calculate allele frequencies and HWE values. Default
is TRUE.}

\item{min.records}{The minimum number of records to perform a test. For
logistic regression, there must be at least this number of each cases and
controls, for linear regression this total number of records. Default is 20}

\item{return.models}{Return a list the complete models, with the names equal
to the string formula used to create them, as well as the results. Default is
FALSE.}

\item{confint.level}{Uses the \code{MASS} package and the \code{confint}
function to calculate a confidence interval at the specified level.
\code{confint} uses a profile likelihood method, which takes some time to
compute. Output is stored in the \code{lower} and \code{upper} columns.
Logistic models will report OR CIs and linear models will report beta CIs.
Default is NA, which does not calculate confidence intervals.}

\item{factor.contrasts}{Contrasts used for factors to generate names used
in clogit.}

\item{my.data}{The data table containing the data for the regression.}

\item{...}{N/A}
}
\value{
n/a
}
\description{
Function to perform and return logistf results on a single genotype-phenotype
association.
}
