% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phecode_exclude}
\alias{phecode_exclude}
\title{Phecode Exclude Table}
\format{
\subsection{\code{phecode_exclude}}{

data frame with 30028 rows of 2 columns
\describe{
\item{code}{the reference code}
\item{exclusion_criteria}{the phecode required to exclude the reference code}
...
}
}
}
\usage{
phecode_exclude
}
\description{
A data map that connects phecodes to their eexclusions
}
\keyword{datasets}
