\name{pheinfo}
\alias{pheinfo}
\docType{data}
\title{
Data frame containing phewas code information.
}
\description{
This data frame contains the description for every mappable phewas code.
}
\usage{data(pheinfo)}
\format{
  A data frame with 1814 observations on the following 2 variables.
  \describe{
    \item{\code{phewas_code}}{a character vector}
    \item{\code{phewas_description}}{a character vector}
  }
}
\source{
http://phewascatalog.org
}
\keyword{datasets}
