\name{phemap}
\alias{phemap}
\docType{data}
\title{
Dataframe containing phewas code mappings.
}
\description{
This data frame maps every ICD9 CM code to its phewas code.
This mapping is many to many.
}
\usage{data(phemap)}
\format{
  A data frame with 20809 observations on the following 2 variables.
  \describe{
    \item{\code{icd9}}{a character vector}
    \item{\code{phewas_code}}{a character vector}
  }
}

\keyword{datasets}
