% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypePlot.R
\name{phenotypePlot}
\alias{phenotypePlot}
\title{Fill in later}
\usage{
phenotypePlot(
  d,
  max.y,
  max.x,
  suggestive.line,
  significant.line,
  size.x.labels = 9,
  size.y.labels = 9,
  switch.axis = F,
  sort.by.value = F,
  sort.by.category.value = F,
  base.labels = F,
  annotate.phenotype.description,
  annotate.angle = 0,
  annotate.size = 5,
  annotate.level,
  annotate.phenotype = F,
  annotate.snp.w.phenotype = F,
  annotate.snp = F,
  annotate.snp.angle = 0,
  annotate.list,
  annotate.only.largest = T,
  lc.labels = F,
  x.group.labels = T,
  x.phenotype.labels = F,
  sizes = F,
  direction = F,
  point.size = 3,
  use.color = T,
  color.palette,
  title = paste0("Phenotype Plot"),
  x.axis.label = "Phenotypes",
  y.axis.label = "Values",
  y.axis.interval = 5
)
}
\arguments{
\item{d}{Data frame containing phenotype and p, or in the case of
phenotypePlot, phenotype and value.}

\item{max.y}{For the Manhattan plots, the maximum -log10(p) on the axis.
Defaults to the larger of 4.4 or the maximum -log10(p) in the supplied data.
For phenotypePlot, the maximum value for the plot. If missing, defaults to
the maximum value in the supplied data.}

\item{max.x}{Maximum number of phenotypes for the plot. If missing, defaults
to the total number.}

\item{suggestive.line}{For phenotypeManhattan, Draw a blue line at this
p-value to show a suggestive significance threshold. Defaults to 0.05.
A missing or NA value as applicable will result in no line. For
phenotypePlot, Blue line drawn at value to emphasize interesting data points.
Missing or an NA value will result in no line.}

\item{significant.line}{For phenotypeManhattan, draw a red line at this
p-value to show an adjusted significance threshold. Defaults to
suggestive.line divided by the number of non-NA p-values. An NA value will
result in no line. For phenotypePlot, the red line is drawn at the specified
value to emphasize interesting data points. Missing or an NA value will
result in no line.}

\item{size.x.labels}{The size of the x axis labels.}

\item{size.y.labels}{The size of the y axis labels.}

\item{switch.axis}{Switch the X and Y axis? Currently non-functioning.
Defaults to false.}

\item{sort.by.value}{Sort the plot by highest to lowest -log10(p) or value?
Defaults to false.}

\item{sort.by.category.value}{Sort the plot by highest to lowest -log10(p)
or value? Defaults to false.}

\item{base.labels}{Use base ggplot2 labels? Defaults to FALSE, which uses
ggrepel.}

\item{annotate.phenotype.description}{Contains either TRUE or a data frame
that contains description annotations for each phenotype. Should contain
columns "phenotype" and "description". Missing or FALSE yields no description
annotation.}

\item{annotate.angle}{Angle for annotation text.}

\item{annotate.size}{Size of annotation tex}

\item{annotate.level}{For phenotypeManhattan or phenotypeManhattan,
annotate points at or below this p with a default of significant.line.
Set to NA to disable adding these annotations. For phenotypePlot, minimum
value to annotate points. Default of no annotation with a missing arugment,
or set to NA to disable.}

\item{annotate.phenotype}{Include the phenotype name in the annotation?

Default is FALSE.}

\item{annotate.snp.w.phenotype}{Include the snp in the phenotype annotation?
Default is FALSE.}

\item{annotate.snp}{Annotate the SNP? This is a second annotation, default is FALSE.}

\item{annotate.snp.angle}{Angle of annotation for the SNP only annotation.}

\item{annotate.list}{A vector of phenotypes to force annotation, regardless of significance.}

\item{annotate.only.largest}{Should only the largest point for each
phenotype be annotated? Default is TRUE, which only shows text for each
phenotype once, even if multiple points pass the annotation threshold or are
listed.}

\item{lc.labels}{Force the labels to lower case? Default is FALSE.}

\item{x.group.labels}{Label the phenotype groups? Requires group attributes
in d. Default is TRUE.}

\item{x.phenotype.labels}{Label every phenotype? Default is no, cannot be
used in conjunction with x.group.labels}

\item{sizes}{Adjust point size based on size in d? Default is FALSE.
Mutually exclusive with OR.size available in the Manhattan plots.}

\item{direction}{Adjust point shape based on direction in d? Default is
FALSE.}

\item{point.size}{Size of the points. Default is 2.}

\item{use.color}{Color the points? Requires color.palette or color in d.
Default is TRUE}

\item{color.palette}{An alternate color palette. Requires at least the
number of groups that exist.}

\item{title}{Title for the plot. Defaults to "Phenotype Plot" with the
date and time.}

\item{x.axis.label}{Title for the plot. Defaults to "Phenotype Plot" with
the date and time.}

\item{y.axis.label}{The label for the y axis. Defaults to a stylized
"-log10(p)" for the Manhattan plots and "Values" for phenotypePlot.}

\item{y.axis.interval}{The interval for y axis labeling. Defaults to 5.
Scale is in -log10 units for the Manhattan plots and is the same as value
for phenotypePlot}
}
\description{
Fill in later
}
