% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_prereq.R
\name{phewas_prereq}
\alias{phewas_prereq}
\title{This is used to prepare and clean phenome and covariates data}
\usage{
phewas_prereq(pheno_dat, demo_dat, number_cases)
}
\arguments{
\item{pheno_dat}{Phenotype data(sparse matrix)}

\item{demo_dat}{Demographics dataframe with id and other covariates}

\item{number_cases}{The criteria of keeping the phecodes with at least how many individual have this code}
}
\description{
This is used to prepare and clean phenome and covariates data
}
