% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewasManhattan.R
\name{phewasManhattan}
\alias{phewasManhattan}
\title{Fill in later}
\usage{
phewasManhattan(
  d,
  annotate.phenotype.description = T,
  pheinfo.map = PheWASmaps::pheinfo,
  ...
)
}
\arguments{
\item{d}{n/a}

\item{annotate.phenotype.description}{n/a}

\item{pheinfo.map}{a map of phecode descriptions for visualization}

\item{...}{n/a}
}
\value{
n/a
}
\description{
Fill in later
}
\examples{

phewasManhattan(PheWAS:::test_phewas, pheinfo.map = PheWASmaps::pheinfo_X)
}
