% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas_plot.R
\name{plot_p}
\alias{plot_p}
\title{Plot P values}
\usage{
plot_p(xwas_result_list, is_fdr = FALSE)
}
\arguments{
\item{xwas_result_list}{list of data frames of EnWAS}

\item{is_fdr}{a flag value decides to plot p-values or FDR}
}
\value{
plot results
}
\description{
Plot P values
}
\examples{
plot_p(list(linear = lm_enwas$enwas_res,
                               ns = ns_enwas$enwas_res))
}
