% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_result.R
\name{result}
\alias{result}
\title{This is used to arrange the result from phewas function}
\usage{
result(result_dat, phecode_description)
}
\arguments{
\item{result_dat}{a dataframe from the result of phewas function}

\item{phecode_description}{a dataframe of the phecode description}
}
\description{
This is used to arrange the result from phewas function
}
