% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data}
\format{
\subsection{\code{sample_data}}{

list of 3
\describe{
\item{id.vocab.code.count}{A data frame containing personID, vocabulary_ID, code, and count}
\item{genotypes}{a dataframe containing the person_ID and rsEXAMPLE status}
\item{id.sex}{a data frame containing the person_ID and their gender}
...
}
}
}
\usage{
sample_data
}
\description{
Sample data for documentation and testing
}
\keyword{datasets}
