% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_phenodat.R
\name{separate_phenome}
\alias{separate_phenome}
\title{Separate large phenome table into chunks by phecodes}
\usage{
separate_phenome(pheno_dat, chunk_number, seed)
}
\arguments{
\item{pheno_dat}{Phenotype data(sparse matrix)}

\item{chunk_number}{Number of chunks}

\item{seed}{The seed for reproducibility}
}
\description{
Separate large phenome table into chunks by phecodes
}
