\name{shiny}
\alias{shiny}
\title{
Provides a plot interface in R Shiny
}
\description{
Loads up a simple R Shiny interface for PheWAS plot manipulation.
}
\usage{
shiny(results)
}

\arguments{
  \item{results}{
The \code{phewas} results object to be plotted.
}
}
\author{
Robert Carroll
}
\note{
This requires the \code{shiny} package.
}

\keyword{ utility }

