% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr_phewas.R
\name{contr.phewas}
\alias{contr.phewas}
\title{Fill in later}
\usage{
contr.phewas(x)
}
\arguments{
\item{x}{A contrasts function used to contrast a column and rename the column}
}
\value{
A data frame with the contrasts and a renamed column name
}
\description{
Fill in later
}
