% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_code_agg.R
\name{default_code_agg}
\alias{default_code_agg}
\title{Sum or count distinct}
\usage{
default_code_agg(index)
}
\arguments{
\item{index}{an index}
}
\value{
an index or a length of an index
}
\description{
FILL IN LATER
}
\examples{
default_code_agg()
}
