% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas.R
\name{enwas}
\alias{enwas}
\title{Environment‐Wide Association Study (EnWAS)}
\usage{
enwas(base_model, exposure_vars, data_set, trans = "none")
}
\arguments{
\item{base_model}{a string of base model}

\item{exposure_vars}{the phenotype list}

\item{data_set}{data set}

\item{trans}{"log" :log and scale transformation or inv: inverse normal transformation}
}
\value{
the model list and EnWAS result
}
\description{
Environment‐Wide Association Study (EnWAS)
}
\examples{
linear_model <- 'BMXWAIST ~ RIDAGEYR*RIAGENDR + BMXBMI'
linear_res <- enwas(linear_model, exposure_vars, data)
}
