% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enwas.R
\name{invNorm}
\alias{invNorm}
\title{Inverse Normal Transformation}
\usage{
invNorm(x, na.last = "keep", ties.method = "average")
}
\arguments{
\item{x}{a numeric, complex, character or logical vector.}

\item{na.last}{for controlling the treatment of NAs. If TRUE, missing values in the data are put last; if FALSE, they are put first; if NA, they are removed; if "keep" they are kept with rank NA.}

\item{ties.method}{a character string specifying how ties are treated, it could be "average", "first", "last", "random", "max", "min"}
}
\value{
transformed data
}
\description{
Inverse Normal Transformation
}
\examples{
invNorm(nhanes$BMXWAIST)
}
