% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinCovar.R
\name{joinCovar}
\alias{joinCovar}
\title{A function to join all the covariate and phenotype files.}
\usage{
joinCovar(pheno, id.sex, covar)
}
\arguments{
\item{pheno}{a wide table with columns representing phecodes and rows
representing people}

\item{id.sex}{an nx2 table with person_ids and gender information}

\item{covar}{a table of variable width consisting of person_IDs and whatever
other information the user requires as covariates or dependent variabes.}
}
\value{
a wide data table consisting of person_ids as rows and covariates and
phecodes as columns
}
\description{
This function takes up to three tables.  These tables include thee phenotype
table (the output of createPhenotype), id.sex (optional), and covar
(data$genotypes in sample data,
but this can be whatever dependent variable and
covariates you're looking for)
}
\examples{
phenotype_data <- createPhenotypes(sample_data$id.vocab.code.count, id.sex = 
sample_data$id.sex)
joinCovar(phenotype_data, sample_data$id.sex, sample_data$genotypes)
}
