% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe_as_ext.R
\name{phe_as_ext}
\alias{phe_as_ext}
\title{Fill in later}
\usage{
phe_as_ext(
  phe.gen,
  additive.genotypes = T,
  min.records = 20,
  return.models = F,
  confint.level = NA,
  factor.contrasts = NA,
  my.data,
  ...
)
}
\arguments{
\item{phe.gen}{N/A}

\item{additive.genotypes}{N/A}

\item{min.records}{N/A}

\item{return.models}{N/A}

\item{confint.level}{N/A}

\item{factor.contrasts}{N/A}

\item{my.data}{N/A}

\item{...}{N/A}
}
\value{
FILL IN LATER
}
\description{
Fill in later
}
\examples{
phe_as_ext('n/a')
}
