% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe_as_logistf.R
\name{phe_as_logistf}
\alias{phe_as_logistf}
\title{To be filled in later}
\usage{
phe_as_logistf(
  phe.gen,
  additive.genotypes = T,
  min.records = 20,
  return.models = F,
  confint.level = NA,
  factor.contrasts = NA,
  my.data,
  ...
)
}
\arguments{
\item{phe.gen}{n/a}

\item{additive.genotypes}{n/a}

\item{min.records}{n/a}

\item{return.models}{n/a}

\item{confint.level}{n/a}

\item{factor.contrasts}{n/a}

\item{my.data}{n/a}

\item{...}{n/a}
}
\value{
n/a
}
\description{
To be filled in later
}
