% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe_as_lrt.R
\name{phe_as_lrt}
\alias{phe_as_lrt}
\title{Function to perform and return lrt results on a single genotype-phenotype
association.}
\usage{
phe_as_lrt(phe.gen, min.records = 20, return.models = T, my.data, ...)
}
\arguments{
\item{phe.gen}{A list with the name of the phecode column being used in the
regression, the column name of independent variable, and covariates}

\item{min.records}{The minimum number of records to perform a test. For
logistic regression, there must be at least this number of each cases and
controls, for linear regression this total number of records. Default is 20}

\item{return.models}{Return a list the complete models, with the names equal
to the string formula used to create them, as well as the results. Default is
FALSE.}

\item{my.data}{The data table containing the data for the regression.}

\item{...}{n/a}
}
\value{
n/a
}
\description{
Function to perform and return lrt results on a single genotype-phenotype
association.
}
