% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe_as_lrt.R
\name{phe_as_lrt}
\alias{phe_as_lrt}
\title{Fill in later}
\usage{
phe_as_lrt(phe.gen, min.records = 20, return.models = T, my.data, ...)
}
\arguments{
\item{phe.gen}{n/a}

\item{min.records}{n/a}

\item{return.models}{n/a}

\item{my.data}{n/a}

\item{...}{n/a}
}
\value{
n/a
}
\description{
Fill in later
}
