% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phecode_map}
\alias{phecode_map}
\title{Map from ICD codes to Phecodes}
\format{
\subsection{\code{phecode_map}}{

a data frame with 84532 rows of 3 columns
\describe{
\item{vocabulary_id}{the ICD source of the ICD code, eg. ICD9CM, ICD10CM}
\item{code}{The ICD code}
\item{phecode}{The applicable phecode}
...
}
}
}
\usage{
phecode_map
}
\description{
A data map that connects icd codes to phecodes
}
\keyword{datasets}
