\name{pheinfo}
\alias{pheinfo}
\docType{data}
\title{
Data frame containing phecode information.
}
\description{
This data frame contains the description and group information for every mappable phecode.
}
\usage{data(pheinfo)}
\format{
  A data frame with 1814 observations on the following 4 variables.
  \describe{
    \item{\code{phecode}}{a character vector}
    \item{\code{phecode_description}}{a character vector}
    \item{\code{phecode_groupnum}}{an integer vector}
    \item{\code{phecode_group}}{a character vector}
  }
}
\source{
http://phewascatalog.org
}
\keyword{datasets}
