% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheinfo}
\alias{pheinfo}
\title{Pheinfo map}
\format{
\subsection{\code{pheinfo}}{

a data frame with 1817 rows of 5 columns
\describe{
\item{phecode}{the reference phecode}
\item{description}{a string describing the phecode}
\item{groupnum}{TBD}
\item{group}{what type of disease the phecode is}
\item{color}{TBD}
...
}
}
}
\usage{
pheinfo
}
\description{
a data map that contains information on each phecode
}
\keyword{datasets}
