\name{pheinfo}
\alias{pheinfo}
\docType{data}
\title{
Data frame containing phecode information.
}
\description{
This data frame contains the description and group information for every mappable phecode.
}
\usage{data(pheinfo)}
\format{
  A data frame with 1817 observations on the following 5 variables.
  \describe{
    \item{\code{phecode}}{a character vector}
    \item{\code{description}}{a character vector}
    \item{\code{groupnum}}{an integer vector}
    \item{\code{group}}{a character vector}
    \item{\code{color}}{a character vector}
  }
}
\source{
http://phewascatalog.org
}
\keyword{datasets}
