\name{phemap}
\alias{phemap}
\docType{data}
\title{
Dataframe containing phecode mappings.
}
\description{
This data frame maps every ICD9 CM code to its phecode.
This mapping is many to many.
}
\usage{data(phemap)}
\format{
  A data frame with 20783 observations on the following 2 variables.
  \describe{
    \item{\code{icd9}}{a character vector}
    \item{\code{phecode}}{a character vector}
  }
}

\keyword{datasets}
