% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeManhattan.R
\name{phenotypeManhattan}
\alias{phenotypeManhattan}
\title{Fill in later}
\usage{
phenotypeManhattan(
  d,
  suggestive.line = 0.05,
  significant.line,
  OR.size = F,
  OR.direction = F,
  sizes,
  annotate.level,
  y.axis.interval = 5,
  y.axis.label = expression(-log[10](italic(p))),
  max.y,
  ...
)
}
\arguments{
\item{d}{n/a}

\item{suggestive.line}{n/a}

\item{significant.line}{n/a}

\item{OR.size}{n/a}

\item{OR.direction}{n/a}

\item{sizes}{n/a}

\item{annotate.level}{n/a}

\item{y.axis.interval}{n/a}

\item{y.axis.label}{n/a}

\item{max.y}{n/a}

\item{...}{n/a}
}
\value{
n/a
}
\description{
Fill in later
}
\examples{
phenotypeManhattan(n/a)
}
