% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypePlot.R
\name{phenotypePlot}
\alias{phenotypePlot}
\title{Fill in later}
\usage{
phenotypePlot(
  d,
  max.y,
  max.x,
  suggestive.line,
  significant.line,
  size.x.labels = 9,
  size.y.labels = 9,
  switch.axis = F,
  sort.by.value = F,
  sort.by.category.value = F,
  base.labels = F,
  annotate.phenotype.description,
  annotate.angle = 0,
  annotate.size = 5,
  annotate.level,
  annotate.phenotype = F,
  annotate.snp.w.phenotype = F,
  annotate.snp = F,
  annotate.snp.angle = 0,
  annotate.list,
  annotate.only.largest = T,
  lc.labels = F,
  x.group.labels = T,
  x.phenotype.labels = F,
  sizes = F,
  direction = F,
  point.size = 3,
  use.color = T,
  color.palette,
  title = paste0("Phenotype Plot ", date()),
  x.axis.label = "Phenotypes",
  y.axis.label = "Values",
  y.axis.interval = 5
)
}
\arguments{
\item{d}{n/a}

\item{max.y}{n/a}

\item{max.x}{n/a}

\item{suggestive.line}{n/a}

\item{significant.line}{n/a}

\item{size.x.labels}{n/a}

\item{size.y.labels}{n/a}

\item{switch.axis}{n/a}

\item{sort.by.value}{n/a}

\item{sort.by.category.value}{n/a}

\item{base.labels}{n/a}

\item{annotate.phenotype.description}{n/a}

\item{annotate.angle}{n/a}

\item{annotate.size}{n/a}

\item{annotate.level}{n/a}

\item{annotate.phenotype}{n/a}

\item{annotate.snp.w.phenotype}{n/a}

\item{annotate.snp}{n/a}

\item{annotate.snp.angle}{n/a}

\item{annotate.list}{n/a}

\item{annotate.only.largest}{n/a}

\item{lc.labels}{n/a}

\item{x.group.labels}{n/a}

\item{x.phenotype.labels}{n/a}

\item{sizes}{n/a}

\item{direction}{n/a}

\item{point.size}{n/a}

\item{use.color}{n/a}

\item{color.palette}{n/a}

\item{title}{n/a}

\item{x.axis.label}{n/a}

\item{y.axis.label}{n/a}

\item{y.axis.interval}{n/a}
}
\value{
n/a
}
\description{
Fill in later
}
\examples{
phenotypePlot(n/a)
}
