% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheSpread.R
\name{pheSpread}
\alias{pheSpread}
\title{A function to spread a long table of phecodes to a wide table}
\usage{
pheSpread(
  phemapped,
  min.code.count = 2,
  add.phecode.exclusions = T,
  id.sex,
  full.population.ids = unique(id.vocab.code.index[[1]]),
  aggregate.fun = default_code_agg,
  vocabulary.map = PheWASmaps::phecode_map,
  rollup.map = PheWASmaps::phecode_rollup_map,
  exclusion.map = PheWASmaps::phecode_exclude,
  gender.exclusion = PheWASmaps::gender_restriction,
  id.name
)
}
\arguments{
\item{phemapped}{a data table of 3 columns: person_ID, phecode, count}

\item{min.code.count}{The minimum code count to be considered a case. NA
results in a continuous output.}

\item{add.phecode.exclusions}{Apply PheWAS exclusions to phecodes.}

\item{id.sex}{If supplied, restrict the phecodes by sex. This should be a
data frame with the first column being the id and the second the sex, "M" or
"F", of the individual. Individuals with any other specification will have
all sex specific phenotypes set to NA.}

\item{full.population.ids}{List of IDs in the "complete" population. This
allows for individuals with no observed codes to have appropriate "control"
status, eg 0s or FALSE in every field}

\item{aggregate.fun}{Aggregate function for duplicated phenotypes
(phecodes, etc) in an individual. The default supports a naive
"distinct date" approach. Use \code{sum} to support count data}

\item{vocabulary.map}{Map between supplied vocabularies and phecodes. Allows
for custom phecode maps. By default uses
\code{\link[PheWASmaps:phecode_map]{PheWASmaps::phecode_map}},
which supports ICD9CM (v1.2) and ICD10CM (beta-2018).
The package also includes the ICD10 beta map
(\code{\link[PheWASmaps:phecode_map_icd10]{PheWASmaps::phecode_map_icd10}}), which
can be used in this parameter.}

\item{rollup.map}{Map between phecodes and all codes that they expand to, eg
parent codes. By default uses the PheWASmaps::phecode_rollup_map.}

\item{exclusion.map}{Map between phecodes and their exclusions. By default
uses the PheWASmaps::phecode_exclude.}

\item{gender.exclusion}{Map determining sex-based exclusions}

\item{id.name}{the name of the phecodes}
}
\value{
A data frame. The first column contains the supplied id for each
individual (preserving the name of the original column). The following
columns are all present phewas codes. They contain T/F/NA for
case/control/exclude or continuous/NA if min.code.count was NA.
}
\description{
The function takes an input of 3 columns - person_ID, phecode, and count.
It returns a wide table with phecodes as TRUE/FALSE/NA. It can
optionally use the PheWAS exclusion criteria.  The function is meant to be
used as a helper function for createPhenotype, but it can be used on its own
as well
}
\examples{
pheSpread(PheWAS:::test_phemapped, full.population.ids=
unique(sample_data$id.vocab.code.index[[1]]), id.name = 
names(sample_data$id.vocab.code.count)[1])
}
