% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewasManhattan.R
\name{phewasManhattan}
\alias{phewasManhattan}
\title{Plotting methods for PheWAS results, phenotype association studies, or
generic phenotype values}
\usage{
phewasManhattan(
  d,
  annotate.phenotype.description = T,
  pheinfo.map = PheWASmaps::pheinfo,
  ...
)
}
\arguments{
\item{d}{Data frame containing phenotype and p, or in the case of
phenotypePlot, phenotype and value}

\item{annotate.phenotype.description}{Contains either TRUE or a data frame
that contains description annotations for each phenotype. Should contain
columns "phenotype" and "description". Missing or FALSE yields no description
annotation.}

\item{pheinfo.map}{a map of phecode descriptions for visualization}

\item{...}{n/a}
}
\value{
A ggplot2 plot. It contains PheWAS codes against their -log10
transformed p-values for phewasManhattan, generic phenotypes versus -log10
transformed p-values for phenotypeManhattan, and generic phenotypes versus
generic values for phenotypePlot.
}
\description{
This functions permit the plotting of a variety of data types found when
working with phenotypes. The function is comprised of sequentially nested
calls as outlined below. Supplying lower level parameters will permit tweaks
to the higher level functions as well.
}
\details{
These functions are nested: phewasManhattan calls phenotypeManhattan which
calls phenotypePlot.

phewasManhattan generates a PheWAS Manhattan plot and can be used with a
result data frame from phewas using phecodes. This function only adds PheWAS
code groupings and descriptions as applicable before calling phenotypeManhattan.
phenotypeManhattan performs the transformations to the data and paramters for
a -log10 scale. It is useful in cases of phewas results using other
phenotypes besides phecodes

phenotypePlot does the actual plotting and can be used to visualize non-p-value data.
All of these functions return ggplot2 objects. Aspects of the plots can be
modified post hoc, for instance to adjust the point size scale one can use
plot + scale_size(\if{html}{\out{<parameters>}})
}
\examples{

phewasManhattan(PheWAS:::test_phewas, pheinfo.map = PheWASmaps::pheinfo_X)
}
