% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewasManhattan.R
\name{phewasManhattan}
\alias{phewasManhattan}
\title{Fill in later}
\usage{
phewasManhattan(d, annotate.phenotype.description = T, ...)
}
\arguments{
\item{d}{n/a}

\item{annotate.phenotype.description}{n/a}

\item{...}{n/a}
}
\value{
n/a
}
\description{
Fill in later
}
\examples{

data <- sample_data
phenotype_data <- createPhenotypes(data$id.vocab.code.count, 
id.sex = data$id.sex) 
final_data <- dplyr::inner_join(dplyr::inner_join(
data$id.sex, data$genotypes),  phenotype_data)
test_phewas <- phewas_ext(names(phenotype_data)[-1], 
                         genotypes = c('rsEXAMPLE'), covariates = 'sex', 
                         data = final_data)
phewasManhattan(test_phewas)
}
