% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkPhewasExport.R
\name{plinkPhewasExport}
\alias{plinkPhewasExport}
\title{Export a table for PheWAS analysis in plink}
\usage{
plinkPhewasExport(phenotypes, file = "plink.pheno", translateIDs = TRUE)
}
\arguments{
\item{phenotypes}{The phenotype file generated by createPhewasTable (or similar).}

\item{file}{The path and filename for the phenotype file.}

\item{translateIDs}{If true, the function will duplicate the first column
(presumably the unique id) into two columns named FID and IID. The default
(TRUE) will handle createPhewasTable input. FALSE requires columns named
"FID" and "IID" to be present.}
}
\value{
a file usable in plink
}
\description{
This function will export a table created by createPhewasTable as a secondary
phenotype file for use in plink.
}
\details{
Plink \url{https://www.cog-genomics.org/plink2} is a commonly used genetics
analysis tool heavily optimized for handling GWAS. This function will export
a phenotype file as expected by plink which can be helpful for analysis.

The \code{translateIDs} options allows users to supply their own FID/IID
pairs in case they are not identical (common in sets with no family information).
}
\examples{
\dontrun{
#Generate some example data
data <- sample_data
#Extract the two parts from the returned list
phenotype_data <- createPhenotypes(data$id.vocab.code.count,
 id.sex = data$id.sex)
 
#Export the file for use in plink
plinkPhewasExport(phenotype_data, file="my.example.pheno")
}
}
